ALFMOB-89: Error screen header not fixed to the top
The divider in the toolbar was not fixed to the top. To resolve this, we now use `frame` and `alignment` to anchor it at the top. However, this also causes views that should be centered within the remaining space to align at the top. To address this, some error views required a `Spacer` at the top and bottom. As part of this fix, a generic `ErrorView was created for use in all error scenarios.